refactor(web): refactor structure of mock handlers
