+++ apps/web/__mocks__/utils/mock-base-url.ts
@@ -1,0 +1,5 @@
+import { env } from '@/env'
+
+const rorBaseApiUrl = env.NEXT_PUBLIC_ROR_API_URL
+
+export const getRorAPIPath = (path: string) => `${rorBaseApiUrl}${path}`
